requirements.txt
flask
python-dotenv
groq
pandas
supabase==2.3.4
gunicorn
